Enable compiler errors and actually write good C

+++ src/async/executor.c
@@ -3,7 +3,7 @@
 void __attribute__((noreturn))
 executor_begin_polling(future_ref futures[], size_t futures_count) {
     while (true) {
-        for (int i = 0; i < futures_count; i++) {
+        for (size_t i = 0; i < futures_count; i++) {
             future_t *future = futures[i];
 
             if (future->ready) {
